requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
multidict==6.0.5
mutagen==1.47.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dotenv==1.0.1
pytube==15.0.0
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
spotipy==2.23.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
youtube-dl==2018.5.30
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz#sha256=076b246aeb4e8018a8675125a1a6389f78724afdd822ba26796b8fadb1662ff2